+++ CHANGELOG.md
@@ -8,5 +8,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Stencil is not enabled in the lighting tab (#1)
+- `Remove Missing Components` not working with child objects (#3)
 
 ## [1.0.0] - 2023-03-13